apart the function
